backend_web/requirements.txt
async-timeout==4.0.3
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
confluent-kafka==2.5.0
Flask==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
numpy==2.0.1
pandas==2.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.8
requests==2.32.3
scikit-learn==1.5.1
scipy==1.14.0
six==1.16.0
threadpoolctl==3.5.0
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3

